% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupAndSort.R
\name{group_and_sort_by_count}
\alias{group_and_sort_by_count}
\title{Group and Sort by Count}
\usage{
group_and_sort_by_count(data, variable_names)
}
\arguments{
\item{data}{the data source to be handled}

\item{variable_names}{variable names which will be used for grouping}
}
\value{
grouped and sorted summarized result
throws error if data is not a dataframe or variable_names has no value
}
\description{
Group the data by variable names and sort them. If there is NA in the value, will consider it as one of the groups.
}
\examples{
library(datateachr)
group_and_sort_by_count(steam_games, genre)
group_and_sort_by_count(vancouver_trees, genus_name)


}
